Removendo interfaces uma vez que optei por usar o wiremock para os requestes de teste
