+++ src/main/java/ifood/component/OpenWeather.java
@@ -1,8 +1,63 @@
 package ifood.component;
 
+import ifood.component.OpenWeather;
+import ifood.exception.BaseException;
+import ifood.exception.ExceptionOriginEnum;
+import ifood.exception.InvalidCityException;
+import ifood.exception.InvalidOpenWeatherResponseException;
 import ifood.model.OpenWeatherResponse;
+import ifood.validator.CityValidator;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
 
-public interface OpenWeather {
+import java.net.URI;
 
-    OpenWeatherResponse getCityTemp(final String cityname, final Double lat, final Double lon);
+@Component
+public class OpenWeather {
+
+    private final String appKey;
+
+    private final RestTemplate restTemplate;
+
+    private final String openWeatherBaseEndpoint;
+
+    @Autowired
+    public OpenWeather(@Value("${openWeather.endpoint}") final String openWeatherEndpoint,
+                       @Value("${openWeather.key}") final String appKey,
+                       final RestTemplate restTemplate) {
+        this.appKey = appKey;
+        this.restTemplate = restTemplate;
+        this.openWeatherBaseEndpoint = openWeatherEndpoint;
+    }
+
+    private URI getWeatherUri(final String cityname, final Double lat, final Double lon) {
+        CityValidator.validate(cityname, lat, lon);
+
+        final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(openWeatherBaseEndpoint);
+        uriBuilder.queryParam("appid", appKey).queryParam("units", "metric");
+
+        return StringUtils.isBlank(cityname)
+                ? uriBuilder.queryParam("lat", lat).queryParam("lon", lon).build().toUri()
+                : uriBuilder.queryParam("q", cityname).build().toUri();
+    }
+
+    public OpenWeatherResponse getCityTemp(final String cityname, final Double lat, final Double lon) {
+        final URI uri = getWeatherUri(cityname, lat, lon);
+        try {
+            return restTemplate.getForObject(uri, OpenWeatherResponse.class);
+        } catch (HttpClientErrorException hcee) {
+            if (HttpStatus.NOT_FOUND.equals(hcee.getStatusCode())) {
+                throw new InvalidCityException(cityname, lat, lon, hcee);
+            }
+            throw new InvalidOpenWeatherResponseException(uri.toString(), hcee);
+        } catch (Exception ex) {
+            throw new BaseException(ex.getMessage(), ex, ExceptionOriginEnum.INTERNAL);
+        }
+    }
 }
