--- src/main/java/ifood/component/impl/OpenWeatherImpl.java
@@ -1,64 +1,0 @@
-package ifood.component.impl;
-
-import ifood.component.OpenWeather;
-import ifood.exception.BaseException;
-import ifood.exception.ExceptionOriginEnum;
-import ifood.exception.InvalidCityException;
-import ifood.exception.InvalidOpenWeatherResponseException;
-import ifood.model.OpenWeatherResponse;
-import ifood.validator.CityValidator;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.URI;
-
-@Component
-public class OpenWeatherImpl implements OpenWeather {
-
-    private final String appKey;
-
-    private final RestTemplate restTemplate;
-
-    private final String openWeatherBaseEndpoint;
-
-    @Autowired
-    public OpenWeatherImpl(@Value("${openWeather.endpoint}") final String openWeatherEndpoint,
-                           @Value("${openWeather.key}") final String appKey,
-                           final RestTemplate restTemplate) {
-        this.appKey = appKey;
-        this.restTemplate = restTemplate;
-        this.openWeatherBaseEndpoint = openWeatherEndpoint;
-    }
-
-    private URI getWeatherUri(final String cityname, final Double lat, final Double lon) {
-        CityValidator.validate(cityname, lat, lon);
-
-        final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(openWeatherBaseEndpoint);
-        uriBuilder.queryParam("appid", appKey).queryParam("units", "metric");
-
-        return StringUtils.isBlank(cityname)
-                ? uriBuilder.queryParam("lat", lat).queryParam("lon", lon).build().toUri()
-                : uriBuilder.queryParam("q", cityname).build().toUri();
-    }
-
-    @Override
-    public OpenWeatherResponse getCityTemp(final String cityname, final Double lat, final Double lon) {
-        final URI uri = getWeatherUri(cityname, lat, lon);
-        try {
-            return restTemplate.getForObject(uri, OpenWeatherResponse.class);
-        } catch (HttpClientErrorException hcee) {
-            if (HttpStatus.NOT_FOUND.equals(hcee.getStatusCode())) {
-                throw new InvalidCityException(cityname, lat, lon, hcee);
-            }
-            throw new InvalidOpenWeatherResponseException(uri.toString(), hcee);
-        } catch (Exception ex) {
-            throw new BaseException(ex.getMessage(), ex, ExceptionOriginEnum.INTERNAL);
-        }
-    }
-}
